Share the conversion code between vulkan and D3D, fixing the missing format thing.
